Changed export method
The plugin export each row as a query, it means that if the exported
database have 10k rows the import tool will execute 10k database queries.
This is not good because make the importing process slow and expensive.

The changes I made it's able to separate each query in 500 rows, it means that
if the database have 10k rows the importing tool will execute only 20 queries.

Beyond less queries the dump now is significantly little compared with
the original dump of this plugin.